rename utils.belts to utils.awards
