+++ dojo_plugin/__init__.py
@@ -19,7 +19,7 @@ from .models import Dojos, DojoChallenges, Belts
 from .config import DOJO_HOST, bootstrap
 from .utils import unserialize_user_flag, render_markdown
 from .utils.discord import get_discord_user, get_discord_roles, add_role, send_message
-from .utils.belts import get_user_belts
+from .utils.awards import get_user_belts
 from .pages.dojos import dojos, dojos_override
 from .pages.dojo import dojo
 from .pages.workspace import workspace
