+++ dojo_plugin/api/v1/belts.py
@@ -1,6 +1,6 @@
 from flask_restx import Namespace, Resource
 
-from ...utils.belts import get_belts
+from ...utils.awards import get_belts
 
 belts_namespace = Namespace("belts", description="Endpoint to manage belts")
